# openapi.yaml
openapi: 3.1.0
...
info:
  title: ine-verificator
  version: "1.0"
  description: "API para verificar documentos oficiales como INE, CSF y documentos bancarios."
  contact:
    name: Soporte Técnico
    email: soporte@ejemplo.com
    url: "https://www.ejemplo.com"
  license:
    name: MIT
    url: "https://opensource.org/licenses/MIT"
servers:
  - url: "http://localhost:3000"
security:
  - Bearer: []
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT # Opcional, pero recomendado para mayor claridad
  schemas:
...
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: Bearer <token>
        id:
          type: string
          description: ID del usuario
          x-stoplight:
            id: c03u3tki1r8mg
...
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - nombre
        - apellido1
        - apellido2
      properties:
        username:
          type: string
          description: Nombre de usuario
        password:
          type: string
          description: Contraseña
        nombre:
          type: string
          description: Nombre
        apellido1:
          type: string
          description: Primer apellido
        apellido2:
          type: string
          description: Segundo apellido
        fecha_nacimiento:
          type: string
          format: date
          description: Fecha de nacimiento
        sexo:
          type: string
          description: Sexo
        direccion:
          type: string
          description: Dirección
    UserDataResponse:
      type: object
      properties:
        nombre:
          type: string
        apellido1:
          type: string
        apellido2:
          type: string
        fecha_nacimiento:
          type: string
          format: date
        sexo:
          type: string
        direccion:
          type: string
        email:
          type: string
          format: email
    UserUpdateRequest:
      type: object
      properties:
        nombre:
          type: string
        apellido1:
          type: string
        apellido2:
          type: string
        fecha_nacimiento:
          type: string
          format: date
        sexo:
          type: string
        direccion:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Mensaje de error
        code:
          type: integer
          example: 400
...
paths:
...
  /login:
    post:
      summary: Login
      description: Inicia sesión y devuelve un Bearer token.
      operationId: loginUser
      tags:
        - Usuarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: Nombre de usuario
                password:
                  type: string
                  description: Contraseña
      responses:
        "200":
          description: Inicio de sesión exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Token expirado o no válido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /register:
    post:
      summary: Registro de Usuario
      description: Registra un nuevo usuario en el sistema.
      operationId: registerUser
      tags:
        - Usuarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Datos de registro inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Usuario ya existe
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /usuario/{usuarioId}:
    get:
      summary: Obtener información del usuario
      description: Devuelve los detalles del usuario autenticado.
      operationId: obtenerDatosUsuario
      tags:
        - Usuarios
      parameters:
        - name: usuarioId
          in: path
          required: true
          schema:
            type: string
          description: ID del usuario a consultar
      responses:
        "200":
          description: Información del usuario obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDataResponse"
        "401":
          description: Token expirado o no válido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Modificar información del usuario
      description: Actualiza los detalles del usuario autenticado.
      operationId: modificarDatosUsuario
      tags:
        - Usuarios
      parameters:
        - name: usuarioId
          in: path
          required: true
          schema:
            type: string
          description: ID del usuario a modificar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: Información del usuario modificada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDataResponse"
        "400":
          description: Datos de modificación inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Token expirado o no válido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"