# verificar-front/src/generated/openapi/model/registerRequest.ts
/**
 * ine-verificator
 * API para verificar documentos oficiales como INE, CSF y documentos bancarios.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: soporte@ejemplo.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterRequest { 
    /**
     * Nombre de usuario
     */
    username: string;
    /**
     * Contraseña
     */
    password: string;
    /**
     * Nombre
     */
    nombre: string;
    /**
     * Primer apellido
     */
    apellido1: string;
    /**
     * Segundo apellido
     */
    apellido2: string;
    /**
     * Fecha de nacimiento
     */
    fecha_nacimiento?: string;
    /**
     * Sexo
     */
    sexo?: string;
    /**
     * Dirección
     */
    direccion?: string;
}