# verificar-front/src/generated/openapi/model/userDataResponse.ts
/**
 * ine-verificator
 * API para verificar documentos oficiales como INE, CSF y documentos bancarios.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: soporte@ejemplo.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserDataResponse { 
    nombre?: string;
    apellido1?: string;
    apellido2?: string;
    fecha_nacimiento?: string;
    sexo?: string;
    direccion?: string;
    email?: string;
}